Add assignin() at end of on_edge_button_press to bring data into workspace --> it works a treat.
Current thoughts:
I was thinking I may need to split 'on_edge_button_press into two scripts - one for left and one for right button?

Presumably Doug found a better way to do this, so that the current_edge data is moved somewhere else --> to the appropriate L/R and 25/50/75 DV level assignment. This would be better.

Look into doing this next and having the checkboxes on the left screen appear checked when the appropriate edge has been added.